Refactor Map into own component, add a routes page, shift Mars map 13 deg
